filer_server main logic is moved to http
